tmux: use 'resize-pane -Z' instead of that crappy tmux-zoom thing
How long has this been built in to tmux without me realizing it? That
tmux-zoom window hack broke the moment you switched windows, but
rezize-pane -Z doesn't actually have to create a new window.

Modify the status bar to put a Z before the window number if it contains
a maximized pane.