tmux.conf: update for version 2.9
According to the tmux 2.9 release notes, "The individual -fg, -bg and
-attr options have been removed; they were superseded by -style options
in tmux 1.9." The vim tmuxline plugin generated this, but apparently
it uses old deprecated options.

Replace everything with -style and we're good.

However, something in 2.9 also screwed up how #[default] works inside
status settings. Seems like it's inheriting the wrong default foreground
color, as a result the inactive windows' status is the wrong font color
and it no longer gets highlighted on background activity.