tmux.conf: don't use 'default' for window status format text
default isn't inheriting from the right place anymore in 2.9, and
I haven't been able to figure out why. Set it explicitly so it's at
least the same text color across the board.

This will kill the background activity color change on installations
with tmux before 2.9, but at least it'll be consistent with my Arch
boxes which all will have 2.9 soon.