vimrc: update airline configuration, fix terminal glitches
It seems like the terminal glitches I've been seeing on Ubuntu 14.04
build servers are from the airline git branch dirty/untracked marker.
I'm not sure why it's broken, or why I only see it on Trusty and not
newer distros, but I don't really care about it and disabling it fixes
the problem.

Additionally, organize the airline config a bit better and do some more
tweaks:
  * enable only git VCS, not other engines
  * hide the "max line number" symbol, it's just unneeded fluff
  * add an autocmd to refresh airline, this is needed to get the git
    branch info to load on the first time a buffer is opened (not sure
    why it's necessary, but it works)