zsh: disable terminal title control for serial consoles
This fixes repeated commands and junk on the serial console on my
Raspberry Pis (and probably any serial console) because minicom doesn't
interpret the escape codes correctly (or something like that).

If the tty is detected as ttyS* or ttyAMA*, then set DISABLE_AUTO_TITLE
to tell oh-my-zsh not to print title info to the terminal between
commands.