zsh: use ZSH_GIT_PROMPT=0 to disable git prompt info
Having gitconfig oh-my-zsh in a particular repo is neat and all, but I'd
rather have an environment variable to set to get the same effect.

Set ZSH_GIT_PROMPT to "0" to disable git status in the prompt (or to any
other value to keep it as usual). Add the function zsh_nogit() to do
this because I'll probably forget that variable name by the next time
i need to use it.