vim: add rust syntax file
It's a lot more up to date than what ships with Vim. This is the version
from commit 0d8ce07aaa3b95e61bf319b25bb3b1a4ecc780c2 of
https://github.com/rust-lang/rust.vim

I could bring in the whole plugin as a submodule, but for now I'm not
interested in all the fancy plugin functionality, just syntax
highlighting.