makefile: use $USERPROFILE
If you change a Windows account name after creating it, the user profile
folder doesn't move, and thus isn't always C:\Users\${USER}. Fortunately
we have the USERPROFILE environment variable set to the correct path, so
we can use that instead.

USERPROFILE contains a windows-style path, so run it through cygpath -p
to convert it to a posix path.