use 'less -Fri' rather than 'smartless' for git pager
It turns out that the broken behavior of 'less -F' without '-X' is fixed
in version 530, from early 2018. -X breaks scrolling and so smartless
was intended to work around that, but it buffers the full output which
has its own problems.

Leave smartless around for easier compatibility on older distros where
I haven't built a newer version from source. Also remove the zsh check,
not sure why I didn't want this for bash.