bin: add tmux-att-or-new script (WIP)
WIP, but the idea is to be able to SSH to a server and immediately
launch/attach tmux. Making this usable will involve some sort of alias
or shell function to run SSH with the right args, and requires that this
script (and python3) be installed on the target server.