myshrc: add vimz, gvimz, and cdz functions; set FZF_DEFAULT_OPTS
vimz, gvimz, and cdz are like the *f call modes of selectf, but
implemented as shell aliases that call fz with the right args.

Omit the 'a' variants and bind ctrl-a to "select all" in fzf. Omit the
'i' variants because fd uses smart-case search by default, and fz passes
arbitrary fd options anyway.