vim: update bitbake files, add a script to help with it
Update bitbake's vim files from the master branch of the bitbake repo,
and add a script to help with that (no submodule because the vim files
are in the contrib directory of the main bitbake repo).

Also keep my patches to the ftdetect plugin, which tweaks when to load
the syntax for .conf and .inc files. Download the new version then apply
a patch embedded in the update-bitbake.sh script.