myshrc.wsl: xlaunch: short sleep between winstart and wsl2-get-display
For some reason, connecting to vcxsrv (via wsl2-get-display) immediately
after launching it causes vcxsrv to crash/exit. It's Windows so I can't
really debug things without digging into procmon or something, but it
only happens on my work computer so I suspect it's something with the
firewall.

Regardless, as janky as it is, putting a short delay between starting
vcxsrv.xlaunch and looking for the server with wsl2-get-display fixes
things.