fz: pass --strip-cwd-prefix to fd
My latest merge of fd 8.6 changes the "cwd prefix" (i.e. './') handling
again. Now it only happens by default in --print0 mode or when directly
executing a command. But for fz I still don't want that, so manually
disable it. This is "safe" enough because we're using print0 wherever
possible already.