vim-latest-release: update for WSL2
Remove github api key. You don't need an API key to get this data (as
long as your IP doesn't hit the rate-limit).

Replace cygstart with wslview to launch the installer. Note that you
can't launch the exe if saved to a Linux path like /tmp, so use wslvar
and wslpath to save to the user's %TEMP% directory when possible.