myshrc: rewrite pathadd, add ~/share/man to MANPATH if it exists
It's been a while since I wrote the pathadd function which depends on
eval at least twice, and the trickiness there was messing with things;
time for a rewrite.

The key improvements are using proper indirect parameter expansion
(${!varname} and ${(P)varname} for bash and zsh, respectively) and
`typeset -g ${varname}=${value}` to set variables, rather than eval.

Other refactoring is just naming local variables better and giving it
the ability to automatically add a trailing colon when manipulating
MANPATH.

Finally, use this to add ~/.share/man to MANPATH when applicable.