zsh: enable async git prompt status
Oh my zsh recently enabled a new feature[1] where it can asynchronously
do the git prompt status stuff, which helps a lot when working in slow
repos. I'm not sure what mechanism is used to redraw the prompt on the
terminal without messing up line editing, but it seems to work okay for
now on zsh 5.9.

Enabling that is done by renaming my version of git_prompt_info to
_omz_git_prompt_info - this is because omz refactored so git_prompt_info
is a wrapper function that registers the async handler for the _omz_
version.

Other cleanup: move DISABLE_UNTRACED_FILES_DIRTY to zwild.zsh instead of
zshrc, and replace one rogue 'git' call with __git_prompt_git.

[1] ec1afe9d feat(git)!: enable async git prompt (now for real)