add support for a ~/myshrc_pre and MYSHRC_PRELOAD_PATHS
Some myshrc functionality checks for the existence of commands (like
bat, rg, and sk) in PATH, but in some environments I have those tools
installed to nonstandard locations. Thus, I need a way to add to PATH
before myshrc gets sourced, rather than the usual afterwards in
myshrc_local.

Add 1) a myshrc_pre file which gets sourced before myshrc (no automatic
upgrade for .bashrc, whatever), and 2) MYSHRC_PRELOAD_PATHS, which is an
array of paths that will be pathadd'd to PATH as soon as pathadd() is
available.