add gitattributes to force text line endings to LF
Cloning this repo on Windows systems (such as for MSYS) shouldn't
check out files as CRLF, since that breaks tmux.conf and possibly other
scripts. This can be worked around by setting core.eol and core.autocrlf
locally in linuxfiles clones on those systems, but it's nicer to force
the right behavior with a gitattributes file in the repo itself.

And re-commit AutoHotkey.ahk as LF instead of CRLF.