vim: add ftplugin/rust.vim override to disable rustfmt formatprg
Earlier this year vim started setting formatprg in rust.vim which causes
text to be formatted with gq to be piped through rustfmt. This doesn't
work when formatting only certain ranges of a file (especially comments)
because rustfmt expects to be given an entire rust source file, not
random snippets.

Add an override file to unset formatprg so that wrapping comments works
again.