vimrc: prevent {<CR> mapping from being used by VSCodeVim
It's been hacky trying to un-map this, since vscode does its own
automatic brace closing, and trying to remap ["{", "<CR>"] to itself
doesn't work well because it still wants to wait for the rest of the
sequence every time I type a {.

There's no way to truly reverse vimrc mappings yet, but I noticed that
if there's leading space on the mapping line, vscode vim's regex won't
match and that line gets skipped.

(note: VSCodeVim runs on the native host and thus uses Windows' vimrc
file, not ~/.vimrc in the wsl remote).